Add files via upload
This is an early game we created at DigiPen meant to mimic the online game Sort the Court. The game will present the user with yes/no prompts and give a random positive or negative outcome to their decisions. The goal is to survive as long as possible. The game is text based and runs in console